Refactor sections into separate component
